clear-cache: exit with success if already clear

+++ clear-cache.sh
@@ -1,2 +1,4 @@
 #!/bin/sh -eux
-rm -v ~/.cache/nix/binary-cache-v6.sqlite*
+if [ -e ~/.cache/nix/binary-cache-v6.sqlite ]; then
+  rm -v ~/.cache/nix/binary-cache-v6.sqlite*
+fi
